Updated bootstrap-modal to use a new ng-view scheme
The body of the modal is loaded from a partial instead of the entire structure.
If the entire structure is loaded from a partial there is an issue with
the close events being bound to the appropriate buttons. This means that you
can't click between views and still close the modal. It also caused issues
with the modal activation button functioning properly. The current way is
slightly less clean and loses a bit of customization but at least it works.

git-svn-id: https://svn.apache.org/repos/asf/oodt/branches/rmcesjpl/projects/rcmes/trunk/rcmet/src/main/ui@1475244 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 4c1e6578a26722da4a11c61634f12ca6eba6e6a6