CLIMATE-583 - Allow for partial evaluation runs
Allow for a user to do partial evaluation runs by specifying that config
parsing errors should be ignored. For instance, suppose that a user
wants to generate only a subregion plot. Normally they would have to
load and run a fake evaluation just to get the subregion plot. With this
new optional argument the user can pass only subregion information and
plot information in the config and still run the evaluation.

Note, there are no guarantees that any partial evaluation will work.
This option gives the user plenty of rope to hang themselves and should
really only be used when you want to generate plots that don't require a
full evaluation run (such as subregion plots).

- Add optional command line arg that tells the parser/runner to ignore
  config parsing errors. This will cause the evaluation to "run" even if
  the configuration isn't value.
- When running an evaluation, only run if the evaluation object passes
  an is_valid check. Normally this is done in the evaluation object
  anyway when calling run but calling this early prevents error messages
  from being dumped.
- Make evaluation creation a bit more robust. Instead of just assuming
  that 'datasets' is always present we first check that the value is not
  None in the config before trying to check for 'reference' or
  'targets'. Similarly with metrics. This is necessary for "partial"
  evaluations.